// Code/SSCMS/src/app/model/raw-material.ts
// src/app/models/raw-material.ts

export interface RawMaterial {
  id?: string;               // auto-generated by json-server
  name: string;             // e.g., "Battery", "Screen"
  description: string;      // optional description
  quantity: number;         // current available units
  unit: string;             // e.g., "pcs", "kg", "liters"
  threshold: number;        // min stock before warning
  supplierId?: number;      // (optional) link to supplier
}